# cpp_info components