--- a/conans/model/build_info.py
+++ b/conans/model/build_info.py
@@ -390,7 +390,7 @@ class DepCppInfo(object):
 
     def as_dict(self):
         result = {}
-        for name in ["name", "rootpath", "sysroot", "system_deps", "libs", "exes",
+        for name in ["name", "rootpath", "sysroot", "description", "system_deps", "libs", "exes",
                      "includedirs", "srcdirs", "libdirs", "bindirs", "builddirs", "resdirs",
                      "include_paths", "src_paths", "lib_paths", "bin_paths", "build_paths", "res_paths",
                      "defines", "cflags", "cxxflags", "cppflags", "sharedlinkflags", "exelinkflags"]: