--- a/conans/model/build_info.py
+++ b/conans/model/build_info.py
@@ -1,5 +1,4 @@
 import os
-from __builtin__ import reversed
 from collections import OrderedDict
 from copy import deepcopy