--- a/conans/model/build_info_components.py
+++ b/conans/model/build_info_components.py
@@ -1,5 +1,4 @@
 import os
-from collections import OrderedDict
 
 from conans.errors import ConanException