Changelog: Feature: Extended `cpp_info` model to support different components inside the same package such as different libraries, executables...
Docs: https://github.com/conan-io/docs/pull/1363

@PYVERS: py34
@TAGS: slow

Closes #5090 
Closes #5242 
Closes #2382
Closes #2387