@lasote 2019-07-17
Test failing on Linux:

```
======================================================================

FAIL: Components can be part of only one configuration

----------------------------------------------------------------------

Traceback (most recent call last):
  File "/tmp/source/0cca/py34/disabled_revisions/conans/test/unittests/model/build_info_test.py", line 649, in components_with_configs_test
    self.assertEqual([], deps_cpp_info["my_lib"].components)
AssertionError: [] != ItemsView(OrderedDict())
```