@memsharded 2019-07-17
> Please put a comment there: "None correspond to the aggregated values from all the dependencies, we can skip it because we are assigning the dependencies individually and can get the aggregated value"

How is the aggregated value computed? In which order are aggregated? Because we don't have the graph information in this file, do we?