@danimtb 2019-07-17
The dependencies are in order in the conanbuildinfo.txt and are added in order to the DepsCppInfo object. The conanbuildinfo.txt is also saved in the same order when that file is generated. The aggregated values respect the order of dependencies and generate the same global values.