@memsharded 2019-07-29
I finally managed to run some performance tests. This branch, merged with develop, against develop.

The resolution of a fully cached and built graph of 200 packages increases with this change from 6.2 seconds to 9.9 seconds. A 100 packages graph from 2.1 to 3.1 seconds. So seems it is 50% added time, linear.